Update Thai VCCV Phonemizer and fix filename type
